T #93 folder name change for old EventService
